# specs/002-title-public-marketplace/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WorkDev Public Marketplace API
  version: 1.0.0
paths:
  /api/public/projects:
    get:
      summary: List public projects
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: skills
          schema:
            type: array
            items:
              type: string
            style: form
        - in: query
          name: budget_min
          schema:
            type: integer
        - in: query
          name: budget_max
          schema:
            type: integer
        - in: query
          name: project_type
          schema:
            type: string
        - in: query
          name: page_size
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectCard'
                  total_count:
                    type: integer
                  next_cursor:
                    type: string

  /api/public/projects/{id}:
    get:
      summary: Get project detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'

  /api/public/developers:
    get:
      summary: List public developers
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: skills
          schema:
            type: array
            items:
              type: string
            style: form
        - in: query
          name: availability
          schema:
            type: string
        - in: query
          name: rate_min
          schema:
            type: integer
        - in: query
          name: rate_max
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeveloperCard'
                  total_count:
                    type: integer
                  next_cursor:
                    type: string

  /api/public/developers/{id}:
    get:
      summary: Get developer detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeveloperDetail'

components:
  schemas:
    ProjectCard:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        short_description:
          type: string
        required_skills:
          type: array
          items:
            type: string
        budget_min:
          type: integer
        budget_max:
          type: integer
        project_type:
          type: string
        client_handle:
          type: string

    ProjectDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        required_skills:
          type: array
          items:
            type: string
        budget_min:
          type: integer
        budget_max:
          type: integer
        project_type:
          type: string
        client:
          type: object
          properties:
            id:
              type: string
            name:
              type: string

    DeveloperCard:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        headline:
          type: string
        primary_skills:
          type: array
          items:
            type: string
        hourly_rate_min:
          type: integer
        hourly_rate_max:
          type: integer

    DeveloperDetail:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        headline:
          type: string
        summary:
          type: string
        portfolio_links:
          type: array
          items:
            type: string
        rating:
          type: number